// qto-ui/src/app/models/disconnect-view.model.ts
import { AbstractBaseModel } from "./abstract-base-model";

export class DisconnectView extends AbstractBaseModel {
  locationId: number;
  orderId: number;
  clientLocationId: number;
  companyName: string;
  companyId: number;
  endCustomerClientId: string;
  parentCompanyName: string;
  parentCompanyId: number;
  parentCompanyClientId: string;
  type: string;
  address: string;
  address1: string;
  address2: string;
  city: string;
  stateProvince: string;
  postalCode: string;
  provisioner: string;
  disconnectReason: string;
  status: string;
  progressPercentage: number;
  provider: string;
  providerOrderSubmitted: Date;
  providerOrderNumber: string;
  customerRequestedDisconnect: Date;
  networkProviderFoc: Date;
  complete: Date;
  created: Date;
  statusAge: number;
  clientServiceId: string;
  mrc: number;
  mrr: number;
  earlyTerminationFee: number;
  billingReviewComplete: Date;
  latestNote: string;
  orderType: string;
  serviceBilledTo: string;
  selected?: boolean;
}